+++ test/golden_test.dart
@@ -1,7 +1,7 @@
 @Tags(<String>['golden'])
 library;
 
-import 'package:fitgrid/fitgrid.dart';
+import 'package:fitgrid_table/fitgrid_table.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';